requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
APScheduler
python-nmap
requests